KonstytucjaRAG/requirements.txt
streamlit
PyPDF2
python-dotenv
langchain
openai
tiktoken
faiss-cpu